refactor : 빈칸 확인 위치 변경
